refactor: introduce @bitwaren/client-type
