+++ .github/CODEOWNERS
@@ -98,6 +98,7 @@ libs/messaging @bitwarden/team-platform-dev
 libs/messaging-internal @bitwarden/team-platform-dev
 libs/serialization @bitwarden/team-platform-dev
 libs/guid @bitwarden/team-platform-dev
+libs/client-type @bitwarden/team-platform-dev
 # Web utils used across app and connectors
 apps/web/src/utils/ @bitwarden/team-platform-dev
 # Web core and shared files
