+++ libs/client-type/README.md
@@ -1,0 +1,5 @@
+# client-type
+
+Owned by: platform
+
+Exports the ClientType enum
