+++ libs/common/src/enums/client-type.enum.ts
@@ -1,10 +1,1 @@
-// FIXME: update to use a const object instead of a typescript enum
-// eslint-disable-next-line @bitwarden/platform/no-enums
-export enum ClientType {
-  Web = "web",
-  Browser = "browser",
-  Desktop = "desktop",
-  // Mobile = "mobile",
-  Cli = "cli",
-  // DirectoryConnector = "connector",
-}
+export { ClientType } from "@bitwarden/client-type";
