+++ libs/common/src/platform/state/user-key-definition.ts
@@ -1,7 +1,7 @@
 // FIXME: Update this file to be type safe and remove this and next line
 // @ts-strict-ignore
-import { array, record } from "@bitwarden/serialization";
 import { isGuid } from "@bitwarden/guid";
+import { array, record } from "@bitwarden/serialization";
 
 import { UserId } from "../../types/guid";
 import { StorageKey } from "../../types/state";
